// monorepo/apps/web/src/modules/warp/packages/client/features/form/types.ts
export type DbAnswersType =
  GetRenderFormDetailsQuery["FormInvitation"][0]["FormSubmissions"][0]["Answers"];

export type DbFormFieldOptionsType =
  GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"];
// ...
export type FormFieldWithChildrenType = FormField & {
  children?: FormFieldWithChildrenType[];
};

export type FormFieldOptionsType =
  | "displayOptions"
  | "interfaceOptions"
  | "fieldOptions";

export type InfoIconPropsType = {
  content: "";
  popoverProps?: PopoverProps;
  popoverDropdownProps?: PopoverDropdownProps;
};

export type FormFieldInterfaces = {
  default: FormFieldWithChildrenType;
  "select-toggle": {
    type: "boolean";
    fieldOptions: {
      label: string;
      enable: boolean;
      readonly: boolean;
      required: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      size: string;
      onLabel: string;
      offLabel: string;
      isHeading?: boolean;
      headingSize?: string;
      showTitleDivider?: boolean;
      infoIconProps?: InfoIconPropsType;
      isAddcomment?: boolean;
      subtitle: string;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-toggle"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "group-wizard": {
    field: string;
    fieldOptions: {
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      template: "wizard1";
      showControls: false;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["group-wizard"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "group-wizard-step": {
    field: string;
    fieldOptions: {
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      title: string;
      infoIconProps?: InfoIconPropsType;
      breadcrumb: string;
      subtitle: string;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["group-wizard-step"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "group-wizard-sub-step": {
    field: string;
    fieldOptions: {
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      title: string;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["group-wizard-sub-step"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "group-tabs": {
    field: string;
    fieldOptions: {
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      template: "pagination";
      showControls: boolean;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["group-tabs"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "select-radio": {
    field: string;
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      choices: { label: string; value: string }[];
      isHeading?: boolean;
      headingSize?: string;
      showTitleDivider?: boolean;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
    };
    display: string;
    displayOptions: {
      orientation: "horizontal" | "vertical";
    };
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-radio"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "select-dropdown": {
    field: string;
    fieldOptions: {
      label: string;
      enable: boolean;
      readonly: boolean;
      required: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      apiUrl: string;
      choices: { label: string; value: string }[];
      defaultValue: string;
      placeholder: string;
      isHeading?: boolean;
      headingSize?: string;
      showTitleDivider?: boolean;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
      maxValue?: number;
    };
    display: string;
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-dropdown"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
// ...
  file: {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      maxSizeInMb: number;
      placeholder?: string;
      accept: string[];
      allowMultiple: boolean;
      directoryPath?: string;
      maxFilesCount: number;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["file"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
// ...
  input: {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder?: string;
      isHeading?: boolean;
      headingSize: string;
      showTitleDivider?: boolean;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
      maxLength?: number;
      maxValue?: number;
      isToolTip: boolean;
      isemail: boolean;
      isAlphabetSpecialChar: boolean;
      isAutoFilled?: boolean;
      isHyperLink?: boolean;
      isAdvance?: boolean;
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["input"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "number-input": {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder?: string;
      isHeading?: boolean;
      headingSize: string;
      showTitleDivider?: boolean;
      numberInputProps?: TextProps;
      mask?: string;
      subtitle: string;
      maxLength?: number;
      maxValue?: number;
      minValue?: number;
      prefix?: string;
      suffix?: string;
      isAutoCalculate?: boolean;
      isToolTip?: boolean;
      decimalLimit?: number;
      isAutoFilled?: boolean;
      isAutoFilledEmail?: boolean;
      isAutoFilledCompanyName?: boolean;
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["number-input"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
    autoCalculatedCalculation?: ValidationRuleType[];
  };
// ...
  "select-multiple-dropdown": {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder: string;
      choices: { label: string; value: string }[];
      creatable: boolean;
      searchable: boolean;
      maxSelectedValues: number;
      isHeading?: boolean;
      headingSize?: string;
      infoIconProps?: InfoIconPropsType;
      isAddcomment?: boolean;
      subtitle: string;
      maxValue?: number;
      description?: string;
      isNumeric?: boolean;
      isDefaultSelected?: true;
      defaultValues?: { key: string; value: string }[];
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-dropdown"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "selected-multiple-dropdown": {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder: string;
      choices: { label: string; value: string }[];
      creatable: boolean;
      searchable: boolean;
      maxSelectedValues: number;
      isHeading?: boolean;
      headingSize?: string;
      infoIconProps?: InfoIconPropsType;
      isAddcomment?: boolean;
      subtitle: string;
      maxValue?: number;
      description?: string;
      isNumeric?: boolean;
      isDefaultSelected?: true;
      defaultValues?: { key: string; value: string }[];
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-dropdown"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
  "select-multiple-checkbox": {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder: string;
      choices: { label: string; value: string }[];
      isHeading?: boolean;
      headingSize?: string;
      showTitleDivider?: boolean;
      selectAllChoice?: string;
      selectNoneChoice?: string;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
    };
    displayOptions: {
      orientation: "horizontal" | "vertical";
    };
    displayRules: RuleType<
      Pick<
        FormFieldInterfaces["select-multiple-checkbox"],
        FormFieldOptionsType
      >
    >[];
    validationRules: ValidationRuleType[];
  };
// ...
  "multi-select-row": {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder: string;
      choices: { label: string; value: string }[];
      creatable: boolean;
      searchable: boolean;
      maxSelectedValues: number;
      isHeading?: boolean;
      headingSize?: string;
      infoIconProps?: InfoIconPropsType;
      isAddcomment?: boolean;
      subtitle: string;
      maxValue?: number;
      description?: string;
      isNumeric?: boolean;
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["select-dropdown"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
  };
// ...
};

export type FormFieldInterfaceKeysType = keyof FormFieldInterfaces;

export type FormFieldInterfaceType<T extends FormFieldInterfaceKeysType> = Omit<
  FormField,
  keyof FormFieldInterfaces[T]
> &
  FormFieldInterfaces[T];

export type RuleType<TOptions = any> = {
  name: string;
  rule: string;
} & TOptions;

export type ValidationRuleType<T = any> = RuleType<
  T & {
    message: string;
  }
>;

export type FormFieldRenderType<T = object> = FC<
  {
    formField: Omit<FormFieldWithChildrenType, keyof T> & T;
  } & SimpleFormFieldControlPropsType
>;

export type GroupWizardPropsType = {
  formField: FormField;
  steps: {
    title: string;
    key: string;
    infoIconProps?: InfoIconPropsType;
    component: JSX.Element;
    breadcrumb: string;
  }[];
};

export type GroupTabsPropsType = {
  formField: FormField;
  tabs: {
    title: string;
    key: string;
    field: string;
    formField: FormFieldWithChildrenType;
  }[];
  steps: {
    title: string;
    key: string;
    infoIconProps?: InfoIconPropsType;
    component: JSX.Element;
    breadcrumb: string;
  }[];
  hierarchyLevel: string;
};

export type SelectTogglePropsType = {
  formField: FormField;
};

export type SelectRadioPropsType = {
  formField: FormField;
};

export type SelectDropdownPropsType = {
  formField: FormFieldInterfaceType<"select-dropdown">;
};

export type GetRenderFormDetailsQueryFormFieldType =
  GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Form"]["FormFields"][0];

export type GetRenderFormDetailsQuerySuggestionType =
  GetRenderFormFieldDetailsQuery["FormInvitation"][0]["Suggestions"];
// ...
export type warningmessageObjectType = {
  isWarningRule?: boolean;
  oldvalue?: string;
  newValue?: string;
  ratio?: string;
  warningmessage?: string;
  formfieldid?: string;
  Threshold_Deviation?: string;
  Deviation_differential?: string;
  questionid?: string;
  ispopupmessageremoved?: boolean;
  isFileUpload?: boolean;
  fileId?: string;
  invitationId?: string;
};